my-jetpack: move connected card to its own folder
