+++ projects/packages/my-jetpack/_inc/components/connected-product-card/index.jsx
