+++ projects/packages/my-jetpack/_inc/components/product-cards-section/boost-card.jsx
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import ProductCard from './connected-product-card';
+import ProductCard from '../connected-product-card';
 
 const BoostCard = ( { admin } ) => {
 	return <ProductCard admin={ admin } slug="boost" />;
